refactor: move deployment document to plugin directory
Since this is really a singleton that's only useful to the deployment tool, it only needs to exist in the plugin's config. Schema import can also be moved to the plugin config.

Added a newDocumentOptions filter to the plugin file to remove the deployment document from the studio's create document list.